move timer runnable to its own class
