+++ analytics/src/main/java/com/bluetriangle/analytics/TrackerExecutor.java
@@ -62,93 +62,4 @@ class TrackerExecutor extends ThreadPoolExecutor {
             super.run();
         }
     }
-
-    /**
-     * A runnable which submits a timer to the given API URL
-     */
-    static class TimerRunnable implements Runnable {
-        /**
-         *  The tracker configuration
-         */
-        final BlueTriangleConfiguration configuration;
-
-        /**
-         * The timer to submit
-         */
-        final Timer timer;
-
-        /**
-         * Create a timer runnable to submit the given timer to the given URL
-         *
-         * @param trackerUrl the tracker API URL to submit the data to
-         * @param timer      the timer to submit
-         */
-        TimerRunnable(@NonNull final BlueTriangleConfiguration configuration, @NonNull final Timer timer) {
-            super();
-            this.configuration = configuration;
-            this.timer = timer;
-        }
-
-        @Override
-        public void run() {
-            HttpsURLConnection connection = null;
-            try {
-                final URL url = new URL(configuration.getTrackerUrl());
-               // Log.d("Tracker URL", String.format("Tracker URL: %s", this.trackerUrl));
-                connection = (HttpsURLConnection) url.openConnection();
-                connection.setRequestMethod("POST");
-                connection.setRequestProperty("Content-Type", "application/json; charset=utf-8");
-                connection.setDoOutput(true);
-                final DataOutputStream dataOutputStream = new DataOutputStream(connection.getOutputStream());
-                dataOutputStream.write(buildBase64EncodedJson());
-                dataOutputStream.close();
-
-                final int statusCode = connection.getResponseCode();
-                if (statusCode >= 300) {
-                    BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(connection.getErrorStream()));
-                    StringBuilder builder = new StringBuilder();
-                    String line = bufferedReader.readLine();
-                    while (line != null) {
-                        builder.append(line);
-                        line = bufferedReader.readLine();
-                    }
-                    bufferedReader.close();
-                    final String responseBody = builder.toString();
-                    configuration.getLogger().error("Server Error submitting %s: %s - %s", timer, statusCode, responseBody);
-                }
-                connection.getHeaderField(0);
-
-            } catch (Exception e) {
-                configuration.getLogger().error(e, "Android Error submitting %s: %s", timer, e.getMessage());
-            } finally {
-                if (connection != null) {
-                    connection.disconnect();
-                }
-            }
-            configuration.getLogger().debug("%s submitted successfully", timer);
-        }
-
-        /**
-         * Builds a JSON string of the timer's fields
-         *
-         * @return a string of JSON representing the timer
-         */
-        private String buildJson() {
-            final JSONObject data = new JSONObject(timer.getFields());
-            //Log.d("crash data sent:", data.toString());
-            return data.toString();
-        }
-
-        /**
-         * Build the base 64 encoded data to POST to the API
-         *
-         * @return base 64 encoded JSON payload
-         * @throws UnsupportedEncodingException if UTF-8 encoding is not supported
-         */
-        private byte[] buildBase64EncodedJson() throws UnsupportedEncodingException {
-            return Base64.encode(buildJson().getBytes("UTF-8"), Base64.DEFAULT);
-        }
-
-    }
-
 }
